Flattened neurons and moved layers to guts
