+++ pkg/networks/guts/sigmoid.go
@@ -1,0 +1,15 @@
+package guts
+
+import (
+	"gonum.org/v1/gonum/mat"
+)
+
+type SigmoidNeuron struct{}
+
+func (neuron *SigmoidNeuron) Evaluate(activations mat.Vector) mat.Vector {
+	panic("implement me")
+}
+
+func (neuron *SigmoidNeuron) Differentiate(activations mat.Vector) mat.Vector {
+	panic("implement me")
+}
