--- pkg/networks/layers/neurons/sigmoid.go
@@ -1,25 +1,0 @@
-package neurons
-
-import (
-	"iio/pkg/vectors"
-)
-
-type SigmoidNeuron struct{
-	neuron *feedforwardNeuron
-}
-
-func (neuron *SigmoidNeuron) FeedForward(activations vectors.Vector) float64 {
-	panic("implement me")
-}
-
-func (neuron *SigmoidNeuron) Derivative(activation float64) float64 {
-	panic("implement me")
-}
-
-func (neuron *SigmoidNeuron) BackPropagate(delta float64) vectors.Vector {
-	panic("implement me")
-}
-
-func (neuron *SigmoidNeuron) Update(bunch *Bunch) {
-	panic("implement me")
-}
