+++ pkg/networks/network.go
@@ -1,7 +1,6 @@
 package networks
 
 import (
-	"iio/pkg/networks/reports"
 	"iio/pkg/sampling"
 	"iio/pkg/vectors"
 )
@@ -9,6 +8,6 @@ import (
 type Network interface {
 	Train(*sampling.Samples)
 	Validate(*sampling.Samples)
-	Test(*sampling.Samples) reports.Report
+	Test(*sampling.Samples) Report
 	Evaluate(vectors.Vector) byte
 }
