+++ pkg/networks/perceptron.go
@@ -1,7 +1,6 @@
 package networks
 
 import (
-	"iio/pkg/networks/reports"
 	"iio/pkg/sampling"
 	"iio/pkg/vectors"
 )
@@ -18,7 +17,7 @@ func (perceptron *Perceptron) Validate(samples *sampling.Samples) {
 	panic("implement me")
 }
 
-func (perceptron *Perceptron) Test(samples *sampling.Samples) reports.Report {
+func (perceptron *Perceptron) Test(samples *sampling.Samples) Report {
 	panic("implement me")
 }
